remove check as nncf now a  required dependency
